+++ CHANGELOG.md
@@ -1,13 +1,22 @@
 # Last changes:
 
 
-Version `v1.0a (12 Mar'25)`
+Version `v1.0b5 (23 Mar'25)`
 
   - BUGFIX: switching "weeks" => \n added to comments
   - BUGFIX: calendar: highlighthing weeks with data (on other years)
   - NEW   : Preferences: week range
   - NEW   : Merge
+  - NEW   : open notes go to report
+  - NEW   : Delete/Backspace on *chk* field clears it; 0/1/2/- sets icon
+  - NEW   : stat.tcl: info on days: data/planned/checked/empty
+  - DELETE: sframe.tcl (not used)
+  - CHANGE: update diagram at exiting AggrEG field
+  - CHANGE: tips on polygons include item name
+  - CHANGE: saving "Y/L/N/Q" i/o "yes/lamp/no/ques" for *chk*
+  - CHANGE: repo.tcl: heading week days: values i/o names
   - CHANGE: tips with comments of calc/chk cells
+  - CHANGE: packages: apave 4.5.7
 
 
 Version `v0.9.6 (8 Mar'25)`